direction int16_t -> int8_t

+++ p_floor.c
@@ -288,7 +288,7 @@ result_e T_MovePlaneFloor(sector_t __far* sector, fixed_t speed, fixed_t dest, i
 }
 
 
-result_e T_MovePlaneCeiling(sector_t __far* sector, fixed_t speed, fixed_t dest, int16_t direction)
+result_e T_MovePlaneCeiling(sector_t __far* sector, fixed_t speed, fixed_t dest, int8_t direction)
 {
   boolean       flag;
   fixed_t       lastpos;
